Add files via upload
DESCRIPTION This code models a aim of this project is to analyze the gym membership data to uncover  patterns, and insights. This will help the gym management improve customer engagement, optimize services, and make data-driven decisions to enhance overall customer satisfaction.

USAGE To use the project clone repository. Run the main.py

CONTRIBUTION This project is open to contribution . To contribute clone this repository.